backend/requirements.txt
fastapi
uvicorn
motor
passlib[bcrypt]
pydantic[email]
python-dotenv
